Add files via upload
gitのURLを提出後に作成したjavaファイル

+++ 20240624/src/Measureable.java
@@ -1,4 +1,4 @@
 interface Measurable {
-    int getArea();
+    double getArea();
     int getPerimeter();
 }
